% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addDataDB}
\alias{addDataDB}
\title{adding  plate data to database}
\usage{
addDataDB(connect_db = NULL, data = NULL, evalPath = "", filename = "",
  startDate = NULL, startTime = NULL, barcode = "", device = "",
  method = "", expName = "", asCSV = FALSE, exportData = FALSE,
  digits = 0, scientific = -3, metaInfo = "", commit = TRUE,
  DB_filename = "/var/local/lara/laraDB.sqlite3")
}
\arguments{
\item{kin_data_df,}{wavelength}
}
\value{
data frame with
}
\description{
Calculates all possible linar models and stores coefficients in a data frame
}
\details{
_____________________________________________________________________________
addDataDB
}
\note{
relativly well runtime optimized version - don't be confused about the ugly code: it's quite fast

todo: add barcode handling !!
}
\examples{
lin_mod_df <- calcAllLinModels(kin_df, wavelength=245)
}
\keyword{plate}
\keyword{readers}

