% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDB_functions.R
\name{addMeasDataDB}
\alias{addMeasDataDB}
\title{addMeasDataDB - adding data to database}
\usage{
addMeasDataDB(connect_db = NULL,
  DB_filename = "/var/local/lara/laraDB.sqlite3", filename = "",
  evalPath = "", barcode = "", device = "", method = "", expName = "",
  startDatetime = NULL, PLC = FALSE)
}
\arguments{
\item{filename=""}{}

\item{method=""}{}

\item{PLC=FALSE}{}
}
\value{
proj_it_id
}
\description{
adding data to database
}
\details{
addMeasDataDB
}
\note{
none

todo: add barcode handling !!
}
\examples{
  lin_mod_df <- calcAllLinModels(kin_df, wavelength=245)
}
\keyword{plate}
\keyword{readers}

