% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addPlateLayoutDB}
\alias{addPlateLayoutDB}
\title{adding container plate layout to database v0.1.0}
\usage{
addPlateLayoutDB(DB_connect = NULL,
  DB_filename = "/var/local/lara/laraDB.sqlite3", barcode = "0000")
}
\arguments{
\item{DB_connect=NULL}{(database connection) - in case one needs to use an exisiting connection}

\item{DB_filename=}{"/var/local/lara/laraDB.sqlite3" (string) - path to SQLite database file}

\item{barcode="0000"}{(string) - barcode for layout file to load}
}
\value{
data frame with
}
\description{
Calculates all possible linar models and stores coefficients in a data frame
}
\details{
_____________________________________________________________________________
addPlateLayoutDB
}
\note{
relativly well runtime optimized version - don't be confused about the ugly code: it's quite fast

todo:
}
\keyword{plate}
\keyword{readers}

