% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDB_functions.R
\name{getEvalParametersDB}
\alias{getEvalParametersDB}
\title{getEvalParametersDB - Get Evaluation Parameter of an Evaluation by ID}
\usage{
getEvalParametersDB(dbConnection = NULL,
  dbFilename = "/var/local/lara/laraDB.sqlite3", evalID = 0)
}
\value{
parameter string
}
\description{
Get Evaluation Parameter of an Evaluation by ID
}
\details{
getEvalParametersDB
}
\note{
none

todo:
}
\keyword{lara_database}
