% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDB_functions.R
\name{getExpDataDB}
\alias{getExpDataDB}
\title{getExpData - Get Data of an Experiment by experiment name and method}
\usage{
getExpDataDB(dbConnection = NULL,
  dbFilename = "/var/local/lara/laraDB.sqlite3", expName = "",
  method = "")
}
\arguments{
\item{method=""}{}
}
\value{
data frame with experiment data
}
\description{
This function retrieved data related to an experiment identified by experiment Name and method name
}
\details{
getExpDataDB
}
\note{
none

todo:
}
\keyword{plate}
\keyword{readers}

