% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDB_functions.R
\name{getExperiments2EvalIDDB}
\alias{getExperiments2EvalIDDB}
\title{getExperiments2EvalIDDB - Get Evaluation Parameters of an Evaluation by IDadding container plate layout to database}
\usage{
getExperiments2EvalIDDB(dbConnection = NULL,
  dbFilename = "/var/local/lara/laraDB.sqlite3", evalID = 0)
}
\value{
data frame with
}
\description{
Get Evaluation Parameters of an Evaluation by IDadding container plate layout to database
}
\details{
getExperiments2EvalIDDB
}
\note{
none

todo:
}
\keyword{plate}
\keyword{readers}
