% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPlateLayoutDB}
\alias{getPlateLayoutDB}
\title{retrieving plate layout from database}
\usage{
getPlateLayoutDB(barcode = "0000", DB_connect = NULL,
  DB_filename = "/var/local/lara/laraDB.sqlite3")
}
\arguments{
\item{DB_connect=NULL}{}

\item{DB_filename="/var/local/lara/laraDB.sqlite3"}{}

\item{barcode="0000"}{}
}
\value{
data frame with
}
\description{
retrieving plate layout from database
}
\details{
getPlateLayoutDB
}
\note{
relativly well runtime optimized version - don't be confused about the ugly code: it's quite fast

todo: add barcode handling !!
}
\examples{
lin_mod_df <- calcAllLinModels(kin_df, wavelength=245)
}
\keyword{plate}
\keyword{readers}

