% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDB_functions.R
\name{addDataDB}
\alias{addDataDB}
\title{addDataDB - adding data to database}
\usage{
addDataDB(connect_db = NULL, data = NULL, evalPath = "", filename = "",
  startDatetime = NULL, barcode = "", device = "", method = "",
  expName = "", asCSV = FALSE, exportData = FALSE, digits = 0,
  scientific = -3, metaInfo = "", commit = TRUE,
  DB_filename = "/var/local/lara/laraDB.sqlite3")
}
\value{
last_data_row
}
\description{
Adding arbitrary data to LARA database
}
\details{
_____________________________________________________________________________
addDataDB
}
\note{
relativly well runtime optimized version - don't be confused about the ugly code: it's quite fast

todo: add barcode handling !!
            add metainfo handling !!
}
\examples{
  addDataDB()
}
\keyword{database}
