% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDB_functions.R
\name{addDataDB}
\alias{addDataDB}
\title{adding  plate data to database}
\usage{
addDataDB(connect_db = NULL, data = NULL, evalPath = "", filename = "",
  startDatetime = NULL, barcode = "", device = "", method = "",
  expName = "", asCSV = FALSE, exportData = FALSE, digits = 0,
  scientific = -3, metaInfo = "", commit = TRUE,
  DB_filename = "/var/local/lara/laraDB.sqlite3")
}
\arguments{
\item{kin_data_df, }{wavelength}
}
\value{
data frame with
}
\description{
Calculates all possible linar models and stores coefficients in a data frame
}
\details{
_____________________________________________________________________________
addDataDB
}
\note{
relativly well runtime optimized version - don't be confused about the ugly code: it's quite fast

todo: add barcode handling !!
}
\examples{
 lin_mod_df <- calcAllLinModels(kin_df, wavelength=245)
}
\keyword{plate}
\keyword{readers}

