% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addMeasDataDB}
\alias{addMeasDataDB}
\title{adding  plate data to database}
\usage{
addMeasDataDB(connect_db = NULL,
  dbFilename = "/var/local/lara/laraDB.sqlite3", filename = "",
  evalPath = "", barcode = "", device = "", method = "", expName = "",
  PLC = FALSE)
}
\arguments{
\item{kin_data_df,}{wavelength}
}
\value{
data frame with
}
\description{
Calculates all possible linar models and stores coefficients in a data frame
}
\details{
addMeasDataDB
}
\note{
relativly well runtime optimized version - don't be confused about the ugly code: it's quite fast

todo: add barcode handling !!
}
\examples{
lin_mod_df <- calcAllLinModels(kin_df, wavelength=245)
}
\keyword{plate}
\keyword{readers}

