% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDB_functions.R
\name{getExpDataByIDDB}
\alias{getExpDataByIDDB}
\title{getExpDataByID - Get Data of an Experiment by ID}
\usage{
getExpDataByIDDB(dbConnection = NULL,
  dbFilename = "/var/local/lara/laraDB.sqlite3", expID = 1, method = "")
}
\arguments{
\item{method=""}{}
}
\value{
data frame with experiment data
}
\description{
This function retrieved data related to an experiment identified by expID and method name
}
\details{
getExpDataByIDDB
}
\note{
none

todo:
}
\keyword{plate}
\keyword{readers}
