% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDB_functions.R
\name{getExperiments2EvalDB}
\alias{getExperiments2EvalDB}
\title{adding container plate layout to database}
\usage{
getExperiments2EvalDB(dbConnection = NULL,
  dbFilename = "/var/local/lara/laraDB.sqlite3", expName = "")
}
\arguments{
\item{kin_data_df, }{wavelength}
}
\value{
data frame with
}
\description{
Calculates all possible linar models and stores coefficients in a data frame
}
\details{
getExperiments2EvalDB
}
\note{
relativly well runtime optimized version - don't be confused about the ugly code: it's quite fast

todo: add barcode handling !!
}
\keyword{plate}
\keyword{readers}

