% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDB_functions.R
\name{getExperiments2EvalDB}
\alias{getExperiments2EvalDB}
\title{getExperiments2EvalDB - Get Evaluation Parameters of an Evaluation by IDadding container plate layout to database}
\usage{
getExperiments2EvalDB(dbConnection = NULL,
  dbFilename = "/var/local/lara/laraDB.sqlite3", expName = "")
}
\value{
data frame with
}
\description{
Get Evaluation Parameters of an Evaluation by IDadding container plate layout to database
}
\details{
getExperiments2EvalDB
}
\note{
none

todo:
}
\keyword{plate}
\keyword{readers}

