% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getExperiments2EvalIDDB}
\alias{getExperiments2EvalIDDB}
\title{adding container plate layout to database}
\usage{
getExperiments2EvalIDDB(dbConnection = NULL,
  dbFilename = "/var/local/lara/laraDB.sqlite3", evalID = 0)
}
\arguments{
\item{kin_data_df,}{wavelength}
}
\value{
data frame with
}
\description{
Calculates all possible linar models and stores coefficients in a data frame
}
\details{
getExperiments2EvalIDDB
}
\note{
relativly well runtime optimized version - don't be confused about the ugly code: it's quite fast

todo: add barcode handling !!
}
\keyword{plate}
\keyword{readers}

